fix(metrics): wire missing fields and resolve CI failures
- Consolidate metrics collection in FrameworkMetrics using shared Arcs.
- Populate all 12 fields in FrameworkMetricsSnapshot with live data.
- Update WASM bindings to expose the complete metrics set.
- Move ChaoticReservoir to src/reservoir_chaotic.rs to satisfy 500 LOC gate.
- Apply cargo fmt to resolve lint warnings.
- Update imports and exports to maintain public API compatibility.

Verified with full test suite and LOC checks.
