+++ src/lib.rs
@@ -101,6 +101,7 @@ mod persistence_versions;
 #[cfg(target_arch = "wasm32")]
 pub mod persistence_wasm;
 pub mod reservoir;
+mod reservoir_chaotic;
 mod reservoir_inertial; // ADR-0064
 mod reservoir_sparse; // LOC gate extraction
 pub mod retrieval;
